+++ src/zProviders/index.ts
@@ -6,6 +6,7 @@ import { ZSymbolProvider } from './zSymbolsProvider';
 import { ZColorProvider } from './zColorProvider';
 import { ZFoldingRangeProvider } from './zFoldingRangeProvider';
 import { ZDocumentLinkProvider } from './zDocumentLinkProvider';
+import { ZHighLightProvider } from './zHighlightProviders';
 
 export { 
     ZHoverProvider, 
@@ -15,5 +16,6 @@ export {
     ZSymbolProvider,
     ZColorProvider,
     ZFoldingRangeProvider,
-    ZDocumentLinkProvider
+    ZDocumentLinkProvider,
+    ZHighLightProvider
 };