+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.0 - 6th March, 2019
+* Implemented [rename](https://code.visualstudio.com/docs/editor/editingevolved#_rename-symbol)
+* Implemented [find references](https://code.visualstudio.com/docs/editor/editingevolved#_peek)
+* Fix an issue when setting a command arg ex (Loop counter) was creating a document variable
+
 ## 1.2.0 - 5th March, 2019
 * Implemented document link provider for `zscriptinsert` command
 * Implemented [workspace symbol providers](https://code.visualstudio.com/docs/editor/editingevolved#_open-symbol-by-name)